pyproject.toml
[tool.poetry]
name = "ml-fairness"
version = "0.1.0"
description = ""
authors = ["christopher-cirelli <ccirelli2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fairlearn = "^0.9.0"
python-decouple = "^3.8"
pre-commit = "^3.5.0"
pylint = "^3.0.2"
pytest = "^7.4.3"
great-expectations = "^0.17.23"
gitpython = "^3.1.40"
mlflow = "^2.7.1"
jupyter = "^1.0.0"
jupyterlab = "^4.0.7"
pendulum = "^2.1.2"
sqlalchemy = "^2.0.23"
pytest-cov = "^4.1.0"
streamlit = "^1.28.1"
black = "^23.11.0"
seaborn = "^0.13.0"
scikit-learn = "^1.3.2"
xgboost = "^2.0.2"
shap = "^0.43.0"
graphviz = "^0.20.1"
lightgbm = "^4.1.0"
hyperopt = "^0.2.7"
pydantic = "1.10.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
